SQL in both of the keep_snapshot code paths must use same number of variables
keep_snapshots was an option we introduced to our fork, and we've now
decided to stop using it - the data storage requirements are significant
and we don't need them any more.

We plan to return to the standard sharejs behaviour of only keeping a
single snapshot per document to optimise the common case of needing the
latest version of a document.

I'm not sure how this discrepancy was introduced, but we havn't noticed
before now because we've had keep_snapshots enabled for as long as we've
used sharejs. I tested setting `keep_snapshots: false` locally in
preparation for the upcoming changes and found the issue.