Revert "Merge pull request #5 from conversation/get-snapshots"
This additional REST endpoint was added to our sharejs fork as a way to
fetch data for the history slider on drafts#edit.

We've since added a /versions REST endpoint and have switched to that
for the same purpose. In the interests of reducing the delta between our
fork and upstream, let's remove the unused endpoint.