Revert "Provide an option to keep historical snapshots in the PostgresQL database adapter."
We added this option to our fork as a way to keep older snapshot data
around so users can see the history of their documents.

These days we don't need the snapshots for that purpose and we have
keep_snapshots set to false, which is the default upstream behaviour.

In the interests of reducing the delta between our fork and upstream,
let's remove the option.