remove keep_operations config option
This option was added to our fork as a way to tackle growth in storage
requirements.

We've since stopped using it (our preference is to keep all operations
so we can retrieve all versions of a document) and in the interests of
reducing the delta between our fork and upstream we can remove it.