Add client side heartbeat check to monitor connection health
Currently, the WebSocket connection can go down without the client
knowing, unless the client sends a frame, it will not initiate a
reconnection.

The heartbeat check should ensure that the client picks up connection
issues more quickly and can reconnect in a more timely manner.

The way this is implemented isn't ideal, but it should get us
across the line while we come up with a long term plan for what we
do about ShareJS in general.