Change the versions API to always return ALL operations
Previously, if there were not enough ops to "fill" an entire
version, they would not be included in the returned array, now
the incomplete version is included in the results.

This does mean each version is no longer guaranteed to return
the same result every time, but in practice this shouldn't be
an issue.